# feat: Complete Phase 1-3 Development Initiative - Stabilization, Documentation, and Performance Benchmarking