--- a/requirements-dev.lock
+++ b/requirements-dev.lock
@@ -1,24 +1,14 @@
-# Development dependency lock file
-# Generated automatically - do not edit manually
-# Python version: 3.12+
-
--e git+https://github.com/docdyhr/versiontracker.git@e90b7ccb824e77754002908a12e9fbbc2a489455#egg=versiontracker
-Authlib==1.6.1
-CacheControl==0.14.3
-Jinja2==3.1.6
-MarkupSafe==3.0.2
-PyYAML==6.0.2
-Pygments==2.19.2
-RapidFuzz==3.13.0
 aiohappyeyeballs==2.6.1
 aiohttp==3.12.14
 aiosignal==1.4.0
 annotated-types==0.7.0
 anyio==4.9.0
 attrs==25.3.0
+Authlib==1.6.1
 bandit==1.8.6
 boolean.py==5.0
 build==1.2.2.post1
+CacheControl==0.14.3
 certifi==2025.7.14
 cffi==1.17.1
 cfgv==3.4.0
@@ -45,10 +35,12 @@ iniconfig==2.1.0
 jaraco.classes==3.4.0
 jaraco.context==6.0.1
 jaraco.functools==4.2.1
+Jinja2==3.1.6
 joblib==1.5.1
 keyring==25.6.0
 license-expression==30.4.4
 markdown-it-py==3.0.0
+MarkupSafe==3.0.2
 marshmallow==4.0.0
 mdurl==0.1.2
 more-itertools==10.7.0
@@ -70,31 +62,33 @@ platformdirs==4.3.8
 pluggy==1.6.0
 pre_commit==4.2.0
 propcache==0.3.2
-psutil==6.1.1
+psutil==7.0.0
 py-serializable==2.1.0
 pycparser==2.22
 pydantic==2.9.2
 pydantic_core==2.23.4
 pydocstyle==6.3.0
+Pygments==2.19.2
 pyparsing==3.2.3
 pyproject_hooks==1.2.0
+pytest==8.4.1
 pytest-asyncio==1.1.0
 pytest-cov==6.2.1
 pytest-mock==3.14.1
 pytest-timeout==2.4.0
 pytest-xdist==3.8.0
-pytest==8.4.1
+PyYAML==6.0.2
+RapidFuzz==3.13.0
 readme_renderer==44.0
 regex==2024.11.6
-requests-toolbelt==1.0.0
 requests==2.32.4
+requests-toolbelt==1.0.0
 rfc3986==2.0.0
 rich==14.0.0
-ruamel.yaml.clib==0.2.12
 ruamel.yaml==0.18.14
+ruamel.yaml.clib==0.2.12
 ruff==0.12.5
 safety-schemas==0.0.14
-safety==3.6.0
 setuptools==80.9.0
 shellingham==1.5.4
 sniffio==1.3.1
@@ -109,14 +103,15 @@ tomlkit==0.13.3
 tqdm==4.67.1
 twine==6.1.0
 typer==0.16.0
-types-PyYAML==6.0.12.20250516
 types-psutil==7.0.0.20250601
+types-PyYAML==6.0.12.20250516
 types-requests==2.32.4.20250611
 types-tabulate==0.9.0.20241207
 types-termcolor==1.1.6.2
 types-tqdm==4.67.0.20250516
 typing_extensions==4.14.1
 urllib3==2.5.0
+-e git+https://github.com/docdyhr/versiontracker.git@64f308d2a34daa14b0c1bac0eaf4b593532c3126#egg=versiontracker
 virtualenv==20.32.0
 wheel==0.45.1
 yarl==1.20.1