--- a/versiontracker/apps/finder.py
+++ b/versiontracker/apps/finder.py
@@ -138,8 +138,8 @@ def get_homebrew_casks_list() -> List[str]:
         # Import the main apps.py file directly (not the apps/ package)
         import importlib.util
         import os
-        
-        _apps_py_path = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'apps.py')
+
+        _apps_py_path = os.path.join(os.path.dirname(os.path.dirname(__file__)), "apps.py")
         _spec = importlib.util.spec_from_file_location("versiontracker_apps_main", _apps_py_path)
         if _spec is not None and _spec.loader is not None:
             _apps_main = importlib.util.module_from_spec(_spec)