… .github/workflows/ci.yml
… .github/workflows/performance.yml
… .github/workflows/release.yml
… .github/workflows/security-audit.yml
… .vscode/extensions.json
… CHANGELOG.md
… CLAUDE.md
… DEVELOPMENT_RECOMMENDATIONS.md
… PHASE_1_3_SUMMARY.md
… TODO.md
… benchmark_results.json
… benchmarks/__init__.py
… benchmarks/core_benchmarks.py
… constraints.txt
… docs/TESTING_STRATEGY.md
… pyproject.toml
… requirements-dev.txt
… requirements.lock
… requirements.txt
… scripts/performance_test.py
… tests/test_additional_functions.py
… tests/test_app_store.py
… tests/test_apps.py
… tests/test_apps_coverage.py
… tests/test_apps_extra.py
… versiontracker/apps/matcher.py
… versiontracker/version/__init__.py
… versiontracker/version/comparator.py
… versiontracker/version/models.py
… versiontracker/version/parser.py